+++ src/main/java/org/kaoden/ws/homework/api/audit/excepition/ExceptionDto.java
@@ -1,0 +1,15 @@
+package org.kaoden.ws.homework.api.audit.excepition;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.experimental.FieldDefaults;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor(staticName = "of")
+@FieldDefaults(level = AccessLevel.PRIVATE)
+public class ExceptionDto {
+    String exceptionMessage;
+}
