--- src/main/java/org/kaoden/ws/homework/controller/entry/exception/ExceptionDTO.java
@@ -1,16 +1,0 @@
-package org.kaoden.ws.homework.controller.entry.exception;
-
-
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.experimental.FieldDefaults;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor(staticName = "of")
-@FieldDefaults(level = AccessLevel.PRIVATE)
-public class ExceptionDTO {
-    String exceptionMessage;
-}
